functiocall

+++ src/lexical.c
@@ -545,7 +545,7 @@ char *yytext;
 	scope_t* scope;
 	uint16_t last_f;
 	bool first_pass_sematic_error_found;
-   uint16_t args_count;
+   int args_count;
    char** args_last_f;
    bool syntax_error_found;
